+++ lg_earth/scripts/kmlsync.py
@@ -53,9 +53,8 @@ def main():
     planet_announce_topic = rospy.get_param('~planet_announce_topic', '/earth/planet')
     pw = PlanetWatcher(planet_announce_topic)
 
-    kml_state = KmlState()
     kmlsync_server.playtour = PlaytourQuery()
-    kmlsync_server.asset_service = rospy.ServiceProxy('/kmlsync/state', kml_state, persistent=False)
+    kmlsync_server.asset_service = rospy.ServiceProxy('/kmlsync/state', KmlState, persistent=False)
     kmlsync_server.playtour_service = rospy.ServiceProxy('/kmlsync/playtour_query', kmlsync_server.playtour, persistent=False)
     kmlsync_server.planet = PlanetQuery()
     kmlsync_server.planet_service = rospy.ServiceProxy('/kmlsync/planet_query', kmlsync_server.planet, persistent=False)